+++ RELEASE_NOTES.md
@@ -48,6 +48,9 @@ Documentation
 Tools
 -----
 
+* IMPROVED: The `lgt2rst.sh` POSIX script now supports passing additional
+options to the `sphinx-quickstart` script when using the `-s` option.
+
 * UPDATED: The `tutor` tool for the refined unknown object warning.
 
 * FIXED: Bug in the `diagrams` tool where a spurious entity node would be